Topologias #1
- Adição de topologias. Classe "enum" definindo os tipos de topologia, mas apenas com "GLOBAL", "RING" e "FOCAL" implementadas
- Adição de recurso para decaimento do coeficiente de inércia
