+++ psoforj/src/main/java/br/com/psoforj/particle/Particle.java
@@ -1,0 +1,40 @@
+package br.com.psoforj.particle;
+
+import br.com.psoforj.util.VectorUtil;
+
+public class Particle {
+	private double[] currentPosition;
+	private double[] bestPosition;
+	private double[] velocity;
+	
+	public Particle(int dimension, double bottomDomainLimit, double topDomainLimit) {
+		this.currentPosition = VectorUtil.generateRandomDoubleVector(dimension, bottomDomainLimit, topDomainLimit);
+		this.bestPosition = this.currentPosition.clone();
+		this.velocity = VectorUtil.generateRandomDoubleVector(dimension, bottomDomainLimit, topDomainLimit);
+	}
+
+	public double[] getCurrentPosition() {
+		return currentPosition;
+	}
+
+	public void setCurrentPosition(double[] currentPosition) {
+		this.currentPosition = currentPosition.clone();
+	}
+	
+	public double[] getBestPosition() {
+		return bestPosition;
+	}
+
+	public void setBestPosition(double[] bestPosition) {
+		this.bestPosition = bestPosition.clone();
+	}
+
+	public double[] getVelocity() {
+		return velocity;
+	}
+
+	public void setVelocity(double[] velocity) {
+		this.velocity = velocity.clone();
+	}
+	
+}
