+++ psoforj/src/main/java/br/com/psoforj/population/TopologyType.java
@@ -1,0 +1,11 @@
+package br.com.psoforj.population;
+
+public enum TopologyType {
+	GLOBAL,
+	RING,
+	MULTI_RING,
+	FOCAL,
+	HIERARCHICAL,
+	VON_NEUMANN,
+	FOUR_CLUSTERS
+}
