+++ psoforj/src/main/java/br/com/psoforj/util/PSOConstants.java
@@ -1,0 +1,5 @@
+package br.com.psoforj.util;
+
+public class PSOConstants {
+	public static final double CLERC_KENNEDY_CONSTRICTION_FACTOR = 0.729;
+}
