--- psoforj/src/main/java/br/com/psoframework/PSOProcess.java
@@ -1,5 +1,0 @@
-package br.com.psoframework;
-
-public class PSOProcess {
-	
-}
